requirements.txt
azure-core==1.35.0
azure-search-documents==11.6.0b12
azure-storage-blob==12.24.1
fastapi==0.116.1
openai==1.107.1
python-multipart==0.0.20
python-dotenv==1.1.1
uvicorn==0.35.0
